+++ scripts/lib/const.js
@@ -1,5 +1,3 @@
-import { MODULE_ID } from "../module.js"
-
 export const CLASS = {
     COLORIZE: {
         PF2E_TOOLBELT: {
@@ -30,5 +28,5 @@ export const QUERY = {
 }
 
 export const TEMPLATES = {
-    RUNES_ON_ITEM: `modules/${MODULE_ID}/templates/runes.hbs`
+    RUNES_ON_ITEM: `modules/sundry/templates/runes.hbs`
 }